IaC/pyproject.toml
[project]
name = "iac"
version = "0.1.0"
description = "Infrastructure as Code components for elrey's LittleLink project"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "aws-cdk-lib>=2.181.1",
    "constructs>=10.4.2",
]

[tool.uv]
python-preference = "only-managed"
